requirements.txt
fastapi
uvicorn[standard]
pdfplumber
pandas
openpyxl
numpy
python-multipart
tqdm
python-dotenv